feat: Add avatar upload
Adds the ability for users to upload their own avatars. Users can now select an image from their computer or device and upload it to their account. The avatar will be displayed in the user's profile and in other places throughout the application.

This feature is useful for users who want to personalize their account and make it more visually appealing.